requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
crispy-bootstrap5==0.7
dj-rest-auth==5.0.1
Django==4.2.4
django-crispy-forms==2.0
django-filter==23.3
django-phonenumber-field==7.1.0
django-rosetta==0.9.9
djangorestframework==3.14.0
drf-spectacular==0.26.5
environs==9.5.0
idna==3.4
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
marshmallow==3.20.1
packaging==23.1
phonenumberslite==8.13.21
Pillow==10.0.0
polib==1.2.0
psycopg2==2.9.7
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
